smart_tts_app/requirements.txt
Django
pyttsx3
numpy
gTTS
edge-tts
python-dotenv
